nomad-next-begin: 3. fetch - suspense
