+++ nomad-next-beginner/next-data-fetching/app/(components)/navigation.tsx
